+++ Mailer/app/Commands/MailWorkerCommand.php
@@ -144,13 +144,13 @@ class MailWorkerCommand extends Command
             if (!$this->mailCache->hasJobs($batch->id)) {
                 $output->writeln("Queue <info>{$batch->id}</info> has no more jobs, cleaning up...");
                 $this->mailCache->removeQueue($batch->id);
-                $this->mailJobBatchRepository->update($batch, ['status' => BatchesRepository::STATUS_DONE]);
+                $this->mailJobBatchRepository->updateStatus($batch, BatchesRepository::STATUS_DONE);
                 $this->mailJobQueueRepository->clearBatch($batch);
                 continue;
             }
 
-            if ($batch->status == BatchesRepository::STATUS_PROCESSED) {
-                $this->mailJobBatchRepository->update($batch, ['status' => BatchesRepository::STATUS_SENDING]);
+            if ($batch->status === BatchesRepository::STATUS_PROCESSED) {
+                $this->mailJobBatchRepository->updateStatus($batch, BatchesRepository::STATUS_SENDING);
             }
 
             $output->writeln("Sending batch <info>{$batch->id}</info>...");
@@ -268,7 +268,7 @@ class MailWorkerCommand extends Command
 
                         if ($this->smtpErrors >= 10) {
                             $this->mailCache->pauseQueue($batch->id);
-                            $this->mailJobBatchRepository->update($batch, ['status' => BatchesRepository::STATUS_WORKER_STOP]);
+                            $this->mailJobBatchRepository->updateStatus($batch, BatchesRepository::STATUS_WORKER_STOP);
                             break;
                         }
                         sleep(10);