+++ Mailer/app/Forms/DennikeTemplateFormFactory.php
@@ -155,7 +155,7 @@ class DennikeTemplateFormFactory
                 $batchStatus = BatchesRepository::STATUS_CREATED;
             }
 
-            $this->batchesRepository->update($batch, ['status' => $batchStatus]);
+            $this->batchesRepository->updateStatus($batch, $batchStatus);
         };
 
         $generate(