+++ Mailer/app/Forms/MediaBriefingTemplateFormFactory.php
@@ -156,7 +156,7 @@ class MediaBriefingTemplateFormFactory
                 $batchStatus = BatchesRepository::STATUS_CREATED;
             }
 
-            $this->batchesRepository->update($batch, ['status' => $batchStatus]);
+            $this->batchesRepository->updateStatus($batch, $batchStatus);
         };
 
         $generate(